Add BKPT fault handling
This change was made based on issue #6 where @dstahlke indicated that
it would be nice to create a dump at a specific location and then
continue execution. I agree!

I added the CRASH_CATCHER_BREAKPOINT() macro to CrashCatcher.h  A
developer can insert this macro in their code to have execution passed
to CrashCatcher, to generate a dump, and then continue execution.

While it is typically safe to continue execution after a breakpoint, it
is often not a good idea to do so for other faults since it can lead to
unpredictable results including an endless loop of dump attempts. A
CrashCatcherInfo structure is now passed into the
CrashCatcher_DumpStart() routine to let the dump handler know if the
stop was caused by a hardcoded BKPT instruction. This will let the
developer know if it is safe to return a CRASH_CATCHER_EXIT code
from their CrashCatcher_DumpEnd() routine or not.

Changes in this commit include:
* Now pass a CrashCatcherInfo structure into CrashCatcher_DumpStart().
  This structure lets the dump handler know 2 things currently:
  * Was the fault caused by a BKPT instruction? It is typically safe
    for a dump handler to return CRASH_CATCHER_EXIT from
    CrashCatcher_DumpEnd() if the cause was BKPT but
    CRASH_CATCHER_TRY_AGAIN should be returned for any other fault
    cause.
  * What was the value of the Stack Pointer, SP, at the time of the
    fault. This could be used in a dump handler to customize the
    RAM regions returned from CrashCatcher_GetMemoryRegions() to
    only dump from the current SP up to the fixed top of stack. This
    would allow for smaller dumps of just the local variables.
* I added a CRASH_CATHCER_BREAKOINT() macro to CrashCatcher.h that
  can be used in developer's code to generate a hardcoded breakpoint.
* CrashCatcher now looks at the instruction which caused the fault to
  see if it is a BKPT instruction. It passes this information along to
  the dump handler in CrashCatcher_DumpStart(). When a breakpoint is
  the cause, CrashCatcher_Entry() advances the PC, Program Counter, to
  the next instruction before returning to HardFault_Handler().
* I updated the assembly language code in HardFault_Handler() to no
  longer infinite loop after CrashCatcher_Entry returns. It now
  restores the non-volatile registers and resumes execution.
* I had to update the unit tests to use a real address for the auto
  stacked value of PC. This PC points to a uint16_t location in memory
  where the machine code for NOP or BKPT can be placed for testing the
  new BKPT handling code.
* HexDump has been modified:
  * The initial text logged now indicates whether a CRASH or BREAKPOINT
    was encountered.
  * Returns CRASH_CATCHER_EXIT from its implementation of
    CrashCatcher_DumpEnd() when the fault cause is BKPT so that
    execution can be continued now that the state at the time of the
    breakpoint has been dumped.
* I updated my CrashingApp sample to now have a test which issues 2
  breakpoints back to back so that I can verify that the registers
  don't get corrupted when I resumed execution from HardFault_Handler.

NOTE: I have only tested this code on a ARMv7-M target, LPC1768, as of
      yet. I will finalize testing on a ARMv6-M target before merging
      into the master branch.